requirements.txt
python-dotenv>=1.0.0

instagrapi>=1.18.0
instaloader>=4.10.0

requests>=2.28.0

# Only if you need advanced concurrency or scheduling:
# aiohttp>=3.8.0